requirements.txt
flask
numpy
scikit-learn
scipy
matplotlib
pandas
open-clip-torch
